Dropping rollup, converting to TSC (just because), chaning toolchain around for tsc, updating exports and package.
