+++ .github/workflows/publish.yaml
@@ -12,6 +12,7 @@ jobs:
         with:
           node-version: '18'
           registry-url: 'https://registry.npmjs.org'
+      - run: git fetch && git pull origin main
       - run: npm ci -w package
       - run: npm run build --if-present -w package
       - run: npm publish --access public -w package
