+++ .gitignore
@@ -1,6 +1,8 @@
 .DS_Store
 .orig
 .npmrc
+*.tsbuildinfo
 
 node_modules
 dist
+build
