+++ demo/package.json
@@ -1,15 +1,12 @@
 {
   "name": "@iwsio/react-blog-examples-demo",
-  "version": "1.0.0",
+  "version": "2.0.0-alpha.1",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4"
+    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "ESLINT_NO_DEV_ERRORS=true BROWSER=none FORCE_COLOR=true react-scripts start | cat -",
@@ -36,5 +33,15 @@
     ]
   },
   "author": "Nathan Bridgewater",
-  "license": "BSD-3-Clause"
+  "license": "BSD-3-Clause",
+  "devDependencies": {
+    "@tailwindcss/typography": "^0.5.8",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "daisyui": "^3.1.6",
+    "postcss": "^8.4.19",
+    "postcss-cli": "^10.0.0",
+    "tailwindcss": "^3.2.4",
+    "web-vitals": "^3.3.2"
+  }
 }
