+++ demo/tailwind.config.js
@@ -1,6 +1,6 @@
 module.exports = {
 	mode: 'jit',
-	content: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html', '../package/src/**/*.{jsx,js}'],
+	content: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html', '../package/src/**/*.{tsx,ts}'],
 	theme: {
 		extend: {}
 	},
