+++ package/package.json
@@ -1,8 +1,9 @@
 {
   "name": "@iwsio/react-blog-examples",
-  "version": "1.0.0",
+  "version": "2.0.0-alpha.1",
   "description": "Example React components for blog posts.",
   "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "exports": {
     ".": "./dist/index.js",
     "./json-csv": "./dist/json-csv/index.js"
@@ -11,9 +12,10 @@
     "dist"
   ],
   "scripts": {
-    "test": "vitest --run",
-    "build": "rollup -c -m",
-    "prebuild": "rimraf dist"
+    "test": "VITEST_SEGFAULT_RETRY=3 vitest --run",
+    "build": "tsc --build tsconfig.build.json",
+    "prebuild": "rimraf dist && tsc --build tsconfig.build.json --clean",
+    "watch": "tsc --build tsconfig.build.json --watch"
   },
   "repository": {
     "type": "git",
@@ -33,37 +35,28 @@
   },
   "homepage": "https://github.com/IWSLLC/react-blog-examples",
   "devDependencies": {
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
-    "@fortawesome/fontawesome-free": "^6.0.0",
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@fortawesome/react-fontawesome": "^0.2.0",
-    "@iwsio/json-csv-core": "^0.2.13",
-    "@rollup/plugin-babel": "^6.0.2",
-    "@rollup/plugin-commonjs": "^23.0.3",
-    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@iwsio/json-csv-core": "^1.0.0",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.0.0",
-    "@testing-library/user-event": "^14.0.4",
-    "eslint": "^8.9.0",
-    "eslint-plugin-import": "^2.25.4",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@typescript-eslint/eslint-plugin": "^5.39.0",
+    "@typescript-eslint/parser": "^5.39.0",
+    "eslint": "^8.10.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-vitest": "^0.0.20",
-    "eslint-plugin-vitest-globals": "^1.2.0",
-    "jsdom": "^20.0.3",
+    "eslint-plugin-react": "^7.29.3",
+    "jsdom": "^22.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "rollup": "^3.3.0",
-    "rollup-plugin-peer-deps-external": "^2.2.4",
-    "vitest": "^0.25.2"
+    "vitest": "^0.32.2"
   },
   "peerDependencies": {
-    "@babel/runtime": "*",
-    "@iwsio/json-csv-core": "*",
-    "react": "^18",
-    "react-dom": "^18"
+    "react": "^18"
+  },
+  "dependencies": {
+    "@iwsio/forms": "^0",
+    "@iwsio/json-csv-core": "^1"
   }
 }
