+++ package/tsconfig.json
@@ -1,0 +1,9 @@
+
+{
+	"extends": "../tsconfig-base.json",
+	"compilerOptions": {
+		"outDir": "./dist",
+		"rootDir": "./src"
+	},
+	"include": ["src/**/*"]
+}
